pyproject.toml
[project]
name = "uptime-kuma-gitops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.10.5",
    "pydantic-settings>=2.7.1",
    "python-socketio[client]>=5.12.1",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
]
